GC2D/ConsoleStr load: 99.5% -> 100% (narrow shine index at call site)

+++ src/GC2D/ConsoleStr.cpp
@@ -85,7 +85,7 @@ void TConsoleStr::load(JSUMemoryInputStream& stream)
 	if (gpMarDirector->mMap != 15) {
 		void* pvVar10 = JKRGetResource("/common/2d/scenarioname.bmg");
 
-		s16 uVar2 = SMS_getShineID(uVar1, uVar9, false);
+		s16 uVar2 = SMS_getShineID(uVar1, (u8)uVar9, false);
 
 		const void* puVar15;
 		if (pvVar10 == nullptr || uVar2 == -1)
